.ctfskel/utils/__init__.py
import requests
import json
import re
import os
import urllib3
from pprint import pprint
from chepy import Chepy

from .ssh import ssh
from .screenshot import screenshot
from .widgets import widgets

# Disable urllib3 insecure warning globally upon import of utils
urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)

__all__ = [
    "requests",
    "json",
    "re",
    "os",
    "urllib3",
    "pprint",
    "Chepy",
    "ssh",
    "screenshot",
    "widgets"
]